Remove generate_testcase binary and move test generation logic to lib.rs
