+++ Cargo.toml
@@ -13,10 +13,6 @@ categories = ["parser-implementations", "text-processing", "web-programming"]
 name = "readability"
 path = "src/main.rs"
 
-[[bin]]
-name = "generate_testcase"
-path = "src/bin/generate_testcase.rs"
-
 [lib]
 name = "readability"
 path = "src/lib.rs"
